Add defensive parsing for EODHD bulk API to handle malformed records
Previously, a single malformed volume value at record #44521 would
cause the entire bulk import to fail. Now records are parsed
individually using JsonDocument, allowing bad records to be skipped
while importing valid ones.

- Changed Volume from long to decimal to handle numeric edge cases
- Added ParseBulkRecordsWithFallback() for record-by-record parsing
- Added GetDecimalOrDefault/GetStringOrDefault for safe field extraction
- Logs skipped record count for monitoring

+++ src/StockAnalyzer.Core/Services/PriceRefreshService.cs
@@ -231,7 +231,7 @@ public class PriceRefreshService : BackgroundService
                     Low = record.Low,
                     Close = record.Close,
                     AdjustedClose = record.AdjustedClose,
-                    Volume = record.Volume
+                    Volume = (long)record.Volume
                 });
                 result.RecordsMatched++;
             }
@@ -398,7 +398,7 @@ public class PriceRefreshService : BackgroundService
                     Low = record.Low,
                     Close = record.Close,
                     AdjustedClose = record.AdjustedClose,
-                    Volume = record.Volume
+                    Volume = (long)record.Volume
                 }).ToList();
 
                 // Insert prices
